Move preview into a separate file
[MAILPOET-2452]
